Move all interop code under Interop namespace
Move all the platform specific interop code to the `Interop` namespace
and also group the P/Invoke calls into classes representing the
modules/libraries/headers the original C functions live in.
