+++ src/shared/Microsoft.Git.CredentialManager/Credential.cs
@@ -7,9 +7,25 @@ using System.Text;
 namespace Microsoft.Git.CredentialManager
 {
     /// <summary>
+    /// Represents a simple credential; user name and password pair.
+    /// </summary>
+    public interface ICredential
+    {
+        /// <summary>
+        /// User name.
+        /// </summary>
+        string UserName { get; }
+
+        /// <summary>
+        /// Password.
+        /// </summary>
+        string Password { get; }
+    }
+
+    /// <summary>
     /// Represents a credential (username/password pair) that Git can use to authenticate to a remote repository.
     /// </summary>
-    public class GitCredential : SecureStorage.ICredential
+    public class GitCredential : ICredential
     {
         public GitCredential(string userName, string password)
         {
@@ -17,22 +33,19 @@ namespace Microsoft.Git.CredentialManager
             Password = password;
         }
 
-        /// <summary>
-        /// User name.
-        /// </summary>
         public string UserName { get; }
 
-        /// <summary>
-        /// Password.
-        /// </summary>
         public string Password { get; }
+    }
 
+    public static class CredentialExtensions
+    {
         /// <summary>
-        /// Returns the base-64 encoded, {username}:{password} formatted string of this `<see cref="GitCredential"/>`.
+        /// Returns the base-64 encoded, {username}:{password} formatted string of this `<see cref="ICredential"/>`.
         /// </summary>
-        public string ToBase64String()
+        public static string ToBase64String(this ICredential credential)
         {
-            string basicAuthValue = string.Format(CultureInfo.InvariantCulture, "{0}:{1}", UserName, Password);
+            string basicAuthValue = string.Format(CultureInfo.InvariantCulture, "{0}:{1}", credential.UserName, credential.Password);
             byte[] authBytes = Encoding.UTF8.GetBytes(basicAuthValue);
             return Convert.ToBase64String(authBytes);
         }
