+++ src/shared/Microsoft.Git.CredentialManager/Interop/Windows/WindowsCredentialManager.cs
@@ -3,9 +3,9 @@
 using System;
 using System.Runtime.InteropServices;
 using System.Text;
-using static Microsoft.Git.CredentialManager.SecureStorage.NativeMethods.Windows;
+using Microsoft.Git.CredentialManager.Interop.Windows.Native;
 
-namespace Microsoft.Git.CredentialManager.SecureStorage
+namespace Microsoft.Git.CredentialManager.Interop.Windows
 {
     public class WindowsCredentialManager : ICredentialStore
     {
@@ -35,27 +35,27 @@ namespace Microsoft.Git.CredentialManager.SecureStorage
 
             try
             {
-                int result = GetLastError(
-                    CredRead(key, CredentialType.Generic, 0, out credPtr)
+                int result = Common.GetLastError(
+                    Advapi32.CredRead(key, CredentialType.Generic, 0, out credPtr)
                 );
 
                 switch (result)
                 {
-                    case OK:
+                    case Common.OK:
                         Win32Credential credential = Marshal.PtrToStructure<Win32Credential>(credPtr);
 
                         var userName = credential.UserName;
 
-                        byte[] passwordBytes = NativeMethods.ToByteArray(credential.CredentialBlob, credential.CredentialBlobSize);
+                        byte[] passwordBytes = InteropUtils.ToByteArray(credential.CredentialBlob, credential.CredentialBlobSize);
                         var password = Encoding.Unicode.GetString(passwordBytes);
 
-                        return new Credential(userName, password);
+                        return new GitCredential(userName, password);
 
-                    case ERROR_NOT_FOUND:
+                    case Common.ERROR_NOT_FOUND:
                         return null;
 
                     default:
-                        ThrowIfError(result, "Failed to read item from store.");
+                        Common.ThrowIfError(result, "Failed to read item from store.");
                         return null;
                 }
             }
@@ -63,7 +63,7 @@ namespace Microsoft.Git.CredentialManager.SecureStorage
             {
                 if (credPtr != IntPtr.Zero)
                 {
-                    CredFree(credPtr);
+                    Advapi32.CredFree(credPtr);
                 }
             }
         }
@@ -87,11 +87,11 @@ namespace Microsoft.Git.CredentialManager.SecureStorage
             {
                 Marshal.Copy(passwordBytes, 0, w32Credential.CredentialBlob, passwordBytes.Length);
 
-                int result = GetLastError(
-                    CredWrite(ref w32Credential, 0)
+                int result = Common.GetLastError(
+                    Advapi32.CredWrite(ref w32Credential, 0)
                 );
 
-                ThrowIfError(result, "Failed to write item to store.");
+                Common.ThrowIfError(result, "Failed to write item to store.");
             }
             finally
             {
@@ -104,20 +104,20 @@ namespace Microsoft.Git.CredentialManager.SecureStorage
 
         public bool Remove(string key)
         {
-            int result = GetLastError(
-                CredDelete(key, CredentialType.Generic, 0)
+            int result = Common.GetLastError(
+                Advapi32.CredDelete(key, CredentialType.Generic, 0)
             );
 
             switch (result)
             {
-                case OK:
+                case Common.OK:
                     return true;
 
-                case ERROR_NOT_FOUND:
+                case Common.ERROR_NOT_FOUND:
                     return false;
 
                 default:
-                    ThrowIfError(result);
+                    Common.ThrowIfError(result);
                     return false;
             }
         }
